blockchain-api/requirements.txt
fastapi
pydantic
web3
uvicorn
SQLAlchemy
mysql-connector-python
flask_mysql_connector
python-dotenv